runtime: regen 2/4 -- explicit-cast drift off master (3bbb04ca4)
44 files (19 flat + 9 windows + 7 linux + 9 darwin, arena.cs x3 included
now that regen 1/4 already cleared their hook lines) picked up an
explicit numeric-type cast the converter now wraps around an arithmetic
expression it previously left unwrapped. Every instance sampled reads
inert: the expression's result was already the target type, so the cast
changes representation, not value.

One shape, sampled verbatim (msize.cs):
  reqSize <= maxSmallSize - mallocHeaderSize
  reqSize <= (uintptr)(maxSmallSize - mallocHeaderSize)

Same shape, a different target type (rwmutex.cs):
  r + 1 == -rwmutexMaxReaders
  r + 1 == (int32)(-rwmutexMaxReaders)

Applied as full-file copies from the fresh L3 multi-platform emission,
not surgical hunks -- verified first that every file in this set had
its ONLY difference in this family (no hook lines, no position-map
content: those live only in package_info.cs, untouched here), so a
whole-file copy is exhaustively equivalent to hand-picking the cast
lines, not a case of "byte-identical-to-emission is a property of the
file, not the change" (G's numstat lesson) -- there is nothing else in
these files to smuggle. Byte-identity to the fresh emission confirmed
per file before staging.

Gates: runtime.csproj builds clean on all three GoTargetOS values
(windows 0/92, linux 0/95, darwin 0/92).

+++ src/core/runtime/map.cs
@@ -831,7 +831,7 @@ search:
             // change those to emptyRest states.
             // It would be nice to make this a separate function, but
             // for loops are not currently inlineable.
-            if (i == abi.MapBucketCount - 1){
+            if (i == (uintptr)(abi.MapBucketCount - 1)){
                 if (b.overflow(Ꮡt) != nil && (~b.overflow(Ꮡt)).tophash[0] != emptyRest) {
                     goto notLast;
                 }
@@ -912,7 +912,7 @@ internal static void mapiterinit(ж<maptype> Ꮡt, ж<hmap> Ꮡh, ж<hiter> Ꮡi
     if (Ꮡh == nil || h.count == 0) {
         return;
     }
-    if (/* unsafe.Sizeof(hiter{}) */ (uintptr)96 / (uintptr)goarch.PtrSize != 12) {
+    if ((uintptr)(/* unsafe.Sizeof(hiter{}) */ (uintptr)96 / (uintptr)goarch.PtrSize) != 12) {
         @throw(hashIterSizeIncorrectˢ); // see cmd/compile/internal/reflectdata/reflect.go
     }
     it.h = Ꮡh;
